requirements.txt
alembic==1.11.3
bleach==6.0.0
blinker==1.6.2
click==8.1.7
dnspython==2.4.2
dominate==2.8.0
email-validator==2.0.0.post2
Faker==12.3.3
Flask==2.3.3
Flask-Bootstrap==3.3.7.1
Flask-Login==0.6.2
Flask-Mail==0.9.1
Flask-Migrate==4.0.4
Flask-Moment==1.0.5
Flask-MySQLdb==1.0.1
Flask-PageDown==0.4.0
Flask-SQLAlchemy==3.0.5
Flask-WTF==1.1.1
greenlet==2.0.2
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.4
Markdown==3.4.4
MarkupSafe==2.1.3
mysqlclient==2.2.0
packaging==23.1
PyJWT==2.8.0
python-dateutil==2.8.2
six==1.16.0
SQLAlchemy==2.0.20
typing_extensions==4.7.1
visitor==0.1.3
webencodings==0.5.1
Werkzeug==2.3.7
WTForms==3.0.1
